refactor(jira-mcp): split app.rs into per-domain modules
Replace the 2055-line monolithic app.rs with crates/jira-mcp/src/app/{mod,
shared,auth,issue,sprint,project,comment,worklog,link,plan,api,request,jql,
notify,tests}.rs. Behavior, public API, and server.rs wiring are unchanged;
the 87 unit tests remain in app/tests.rs. Groundwork for the upcoming
watcher, jql_build, attachment, and board tool additions.
